Reviews update
new icon for the tabs, all say "- FMP" for the tabs, added "click to change" for star rating section so it's easier to interpret, and formatted the average rating part. Needs to display review average and total reviews still. The single star should stay fully filled in, it's just to indicate that that is the star rating.